pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "desloppify"
version = "0.2.0"
description = "Multi-language codebase health scanner and technical debt tracker"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = ["Pillow>=9.0.0"]

[project.scripts]
desloppify = "desloppify.cli:main"

[tool.setuptools.packages.find]
include = ["desloppify*"]
